// template/apps/api/src/addresses/entities/address.entity.ts
import { ApiProperty } from '@nestjs/swagger';
import { Address } from '@workspace/database';

export class AddressEntity implements Address {
  /*
  this code was not generated by prisma-generator-nestjs-dto
  but manually added to demonstrate usage of decorators along with @nestjs/swagger Exluce/Expose Properties
  the reason to implement an entity like this, is to be able to exlude certain properties from the response 
  and to document it accordingly with the @nestjs/swagger @ApiReponse decorators
  e.g. password fields, createdAt, updatedAt, etc.
  an example is found in this prisma-nestjs documentation: https://www.prisma.io/blog/nestjs-prisma-relational-data-7D056s1kOabc#generate-new-users-rest-resource
  */
  constructor(partial: Partial<AddressEntity>) {
    Object.assign(this, partial);
  }

  @ApiProperty({
    type: 'string',
  })
  addressId: string;
  @ApiProperty({
    description: 'Subdomain of the tenant',
    minimum: 5,
    type: 'string',
    nullable: false,
  })
  city: string;
  @ApiProperty({
    description: 'Country of the address',
    minimum: 5,
    type: 'string',
    nullable: false,
  })
  country: string;
  @ApiProperty({
    description: 'Post code of the address',
    minimum: 5,
    type: 'string',
    nullable: false,
  })
  postCode: string;
  @ApiProperty({
    description: 'State of the address',
    minimum: 5,
    type: 'string',
    nullable: false,
  })
  state: string;
  @ApiProperty({
    description: 'Street name of the address',
    minimum: 5,
    type: 'string',
    nullable: false,
  })
  streetName: string;
  @ApiProperty({
    description: 'Street number of the address',
    minimum: 5,
    type: 'string',
    nullable: false,
  })
  streetNumber: string;
  @ApiProperty({
    description: 'Creation date of the address',
    type: 'string',
    required: false,
    nullable: true,
  })
  createdAt: Date;
  @ApiProperty({
    description: 'Last updated date of the address',
    type: 'string',
    required: false,
    nullable: true,
  })
  modifiedAt: Date;
  @ApiProperty({
    description: 'Last updated date of the address',
    type: 'string',
    required: false,
    nullable: true,
  })
  updatedAt: Date;
  @ApiProperty({
    description: 'Indicates if the address is deleted',
    type: 'boolean',
    required: false,
    nullable: false,
  })
  deleted: boolean;
  @ApiProperty({
    description: 'Longitude of the address',
    type: 'number',
    required: false,
    nullable: false,
  })
  longitude: number | null;
  @ApiProperty({
    description: 'Latitude of the address',
    type: 'number',
    required: false,
    nullable: false,
  })
  latitude: number | null;
  @ApiProperty({
    description: 'ID of the tenant associated with the address',
    type: 'string',
    required: false,
    nullable: false,
  })
  tenantId: string;
}